q2/requirements.txt
streamlit==1.31.1
sentence-transformers==2.5.1
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.1
openai==1.12.0
seaborn==0.13.2
matplotlib==3.8.3
python-dotenv==1.0.1 